requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
requests
jinja2
groq